# Improve Minisat::vec<T> implementation and usage