… minisat/mtl/Vec.h
… minisat/simp/SimpSolver.cc