These two commits first improve the internals of `Minisat::vec<T>` and then how Minisat itself uses the vector. For more explanations and details see the commit message of the first commit.


## Measurements

I did some informal measurements and the performance impact of these changes is either tiny bit beneficial, or imperceivable. On the set of benchmark in the repository, there were no performance regressions.