@horenmar 2018-09-11
I cleaned up the PR a bit and after the CI runs, I am going to merge it.