@horenmar 2018-09-11
The AppVeyor failures are spurious (it sometimes fails to clone the repo), so I am going to merge it anyway.